Threading semi-working
Got the threading to mostly work. I need to separate the new data from
the old data though as I'm saving the old data over and over again.

I'm also about to delete all the _all stuff. That's just an ok feature,
and it's a pain to deal with. I can re-implement it later if desired.